moved helper functions to bddutils somewhere in the converter hierarchy
